Prevent exposure of internal service ports
Both the containers patient-service-db and auth-service-db running the postgres database,
are now configured to only expose their ports to the internal Docker network, preventing external access.
Internal Container Port: The PostgreSQL image (postgres:latest) by default runs its database server on port 5432
inside the container. This is its default listening port.
Docker Compose Internal Networking:
You have defined a networks: section for both auth-postgres and auth-service, placing them both on the internal network.
When containers are on the same Docker network (like internal in your case), they can communicate with each other directly using their service names as hostnames. Docker's internal DNS handles this resolution.
Crucially, when they communicate within the same Docker network, they do so via the internal ports of the containers.

ports Mapping is for Host-to-Container:
The ports: - "5433:5432" mapping on auth-postgres means: "Map port 5433 on the host machine to port 5432 inside the auth-postgres container."
This mapping is only necessary if you want to access the database from your host machine (e.g., using localhost:5433 in a database client) or from other external networks.
It is not needed for communication between containers within the same Docker network. Only (like docker exec -it <container_name> bash)
we need to access the database from the host machine that use the 5433 port mapping.

The pgadmin which is also running in a container, is configured to connect to the auth-postgres service using the internal Docker network.
This means it will use the service name auth-postgres as the hostname and port 5432
for its connection, ensuring that it communicates over the internal network without exposing the database to the outside world.